requirements.txt
alabaster==0.7.12
anyio==3.6.2
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
astunparse==1.6.3
attrs==22.2.0
Babel==2.11.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
colorama==0.4.6
cycler==0.11.0
debugpy==1.6.4
decorator==5.1.1
defusedxml==0.7.1
docopt==0.6.2
docutils==0.19
entrypoints==0.4
fastjsonschema==2.16.2
fonttools==4.38.0
idna==3.4
imagesize==1.4.1
importlib-metadata==5.2.0
importlib-resources==5.10.1
ipykernel==6.16.2
ipython==7.34.0
ipython-genutils==0.2.0
jedi==0.18.2
Jinja2==3.1.2
jittor==1.3.6.5
json5==0.9.10
jsonschema==4.17.3
jupyter-server==1.23.4
jupyter_client==7.4.8
jupyter_core==4.12.0
jupyterlab==3.5.2
jupyterlab-pygments==0.2.2
jupyterlab_server==2.16.6
kiwisolver==1.4.4
livereload==2.6.3
MarkupSafe==2.1.1
matplotlib==3.5.3
matplotlib-inline==0.1.6
maxflow==0.0.1
mistune==2.0.4
nbclassic==0.4.8
nbclient==0.7.2
nbconvert==7.2.7
nbformat==5.7.1
nest-asyncio==1.5.6
notebook==6.5.2
notebook_shim==0.2.2
numpy==1.26.2
opencv-python==4.8.1.78
packaging==22.0
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.3.0
pkgutil_resolve_name==1.3.10
prometheus-client==0.15.0
prompt-toolkit==3.0.36
psutil==5.9.4
ptyprocess==0.7.0
pycparser==2.21
Pygments==2.13.0
PyMaxflow==1.3.0
pyparsing==3.0.9
pyrsistent==0.19.2
python-dateutil==2.8.2
pytz==2022.7
pyzmq==24.0.1
requests==2.28.1
scipy==1.11.4
Send2Trash==1.8.0
six==1.16.0
sniffio==1.3.0
snowballstemmer==2.2.0
soupsieve==2.3.2.post1
Sphinx==5.3.0
sphinx-autobuild==2021.3.14
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
terminado==0.17.1
tinycss2==1.2.1
tomli==2.0.1
tornado==6.2
tqdm==4.64.1
traitlets==5.8.0
typing_extensions==4.4.0
urllib3==1.26.13
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.4.2
yarg==0.1.9
zipp==3.11.0
